Added binding between ExperienceSystem and Player UI to correctly update XP bar on kill and level up

+++ Source/Mortalis/GameFramework/Enemies/EnemyBase.cpp
@@ -34,7 +34,7 @@ void AEnemyBase::ProcessDeath()
 	}
 }
 
-int32 AEnemyBase::GetExperience() const
+float AEnemyBase::GetExperience() const
 {
 	return Experience;
 }
